Changed File Structure
Integrated record.py and transcript.py into input_pipeline.py, created a core directory for the core parts of melody
